fix(ci): satisfy platform ruff + mcp-manager golangci-lint
- test_file_toolset: drop redundant mid-file imports (E402/F821), use top-level
- tool_authorization: sort __all__ (RUF022)
- activation-service: remove dead hydrationContainsPath/hydrateWorkspace (unused),
  fmt.Fprintf instead of Write([]byte(Sprintf)) (QF1012)
- goimports on workspace repository{,_test}.go; drop unnecessary int64() (unconvert)
- validateExecutionResult: //nolint:gocyclo (branch's documented complexity-debt scope)

+++ agentarea-mcp-manager/internal/sandboxcontrol/transport_test.go
@@ -222,7 +222,7 @@ func createRefsOnlyExecution(t *testing.T, ctx context.Context, service *Service
 			ObjectURI:           fmt.Sprintf("s3://trusted/workspaces/workspace-canary/tasks/task-%d/objects/%s", generation, digest),
 			ObjectVersionOrETag: "etag-immutable",
 			SHA256:              digest,
-			Size:                int64(objectSize),
+			Size:                objectSize,
 		}},
 	})
 	if err != nil {
